requirements.txt
Pillow
geopy
requests
